requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.6
defusedxml==0.6.0
Django==2.2.5
django-allauth==0.44.0
django-appconf==1.0.4
django-cors-headers==3.7.0
django-imagekit==4.0.2
django-money==1.3.1
django-rest-auth==0.9.5
django-summernote==0.8.11.6
djangorestframework==3.12.1
djangorestframework-api-key==2.0.0
gunicorn==20.0.4
idna==2.10
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.1
pilkit==2.0
Pillow==8.1.0
psycopg2==2.8.6
py-moneyed==0.8.0
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
whitenoise==5.2.0
